requirements.txt
aiohttp==2.2.5
async-timeout==2.0.1
beautifulsoup4==4.6.0
certifi==2017.11.5
chardet==3.0.4
cleverwrap==0.2.3.6
discord.py==1.0.0a0
Django==2.0
django-colorfield==0.1.14
docopt==0.6.2
facebook-sdk==2.0.0
gevent==1.2.2
greenlet==0.4.12
holster==1.0.16
html5lib==1.0.1
idna==2.6
multidict==4.1.0
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.4
pyaes==1.6.1
pyasn1==0.4.2
pytz==2017.3
requests==2.13.0
rsa==3.4.2
six==1.10.0
Telethon==0.17.4.3
twitter==1.18.0
urllib3==1.22
webencodings==0.5.1
websocket-client==0.40.0
websockets==4.0.1
yarg==0.1.9
yarl==1.1.1
